// src/main/resources/public/ts/types/Factory.d.ts
import { Meta64 } from "./Meta64";
import { SystemFolder } from "./SystemFolder";
import { Attachment } from "./Attachment";
import { Edit } from "./Edit";
import { Encryption } from "./Encryption";
import { DomBind } from "./DomBind";
import { Nav } from "./Nav";
import { Podcast } from "./Podcast";
import { Props } from "./Props";
import { Render } from "./Render";
import { Search } from "./Search";
import { Share } from "./Share";
import { Tag } from "./Tag";
import { User } from "./User";
import { Util } from "./Util";
import { View } from "./View";
import { RssPlugin } from "./plugins/RssPlugin";
import { CoreTypesPlugin } from "./plugins/CoreTypesPlugin";
export declare class Factory {
    private meta64;
    private tag;
    private util;
    private edit;
    private attachment;
    private encryption;
    private nav;
    private props;
    private render;
    private srch;
    private share;
    private user;
    private view;
    private podcast;
    private systemFolder;
    private domBind;
    private rssPlugin;
    private coreTypesPlugin;
    constructAll: () => void;
    postConstructAll: () => void;
    getMeta64: () => Meta64;
    getTag: () => Tag;
    getUtil: () => Util;
    getEdit: () => Edit;
    getAttachment: () => Attachment;
    getEncryption: () => Encryption;
    getNav: () => Nav;
    getProps: () => Props;
    getRender: () => Render;
    getSearch: () => Search;
    getShare: () => Share;
    getUser: () => User;
    getView: () => View;
    getPodcast: () => Podcast;
    getSystemFolder: () => SystemFolder;
    getDomBind: () => DomBind;
    getRssPlugin: () => RssPlugin;
    getCoreTypesPlugin: () => CoreTypesPlugin;
}